+++ R/fct_integration.R
@@ -34,6 +34,22 @@ dMeasureIntegration <- function(information) {
   if (information == "moduleID") {return(c("Appointments_dt"))}
 }
 
+#' sidebarmenuPriority
+#'
+#' @name sidebarmenuPriority
+#'
+#' @description priority order to appear in sidebarmenu
+#'   integration with DailyMeasure
+#'
+#' 50 is medium priority. larger numbers are higher priority
+#'
+#' @return priority.
+#'
+#' @export
+sidebarmenuPriority <- function() {
+  return(99) # very high priority
+}
+
 #' dMeasureAppointments class
 #' @title dMeasureAppointments class
 #' @description list appointments by clinician providers