+++ NEWS.md
@@ -5,6 +5,21 @@ date: "09/10/2021"
 output: html_document
 ---
 
+# dMeasureAppointments 1.0.0
+10th October 2021
+
+* working release
+
+## New
+
+* add `sidebarmenuPriority` to `fct_integration.R`
+  - allows 'high' priority in sidebar menu display
+  
+## Changes
+
+* removed 'Print view' - which was not being used
+
 # dMeasureAppointments 0.0.0.9000
+9th October 2021
 
 * Added a `NEWS.md` file to track changes to the package.